images are finally being deleted

+++ routes/product.js
@@ -4,6 +4,7 @@ const mongo = require("mongoose");
 const scrapeProduct = require("../scrape");
 const fs = require('fs')
 let mailer = require("nodemailer");
+const { type } = require("os");
 
 // mongo schemas
 const Product = require("../modals/productSchema").Product;
@@ -29,13 +30,20 @@ router.get("/notify", (req, res) => {
 });
 
 router.delete("/deleteImg", (req, res) => {
-    const key = req.query.key;
-    const path = `./frontend/public/screenshots/${key}.png`;
+    const keys = req.query.key.split(`"`);
+    let path;
 
     try {
-        // removed file
-        fs.unlinkSync(path);
-        console.log(`deleted image: ${path}.png`)
+        keys.forEach(key => {
+            if (key.length === 8) {
+                // remove file
+                path = `./frontend/public/screenshots/${key}.png`;
+                fs.unlinkSync(path);
+                console.log(`deleted image: ${path}.png`);
+            }
+        });
+
+        res.json(null);
     } catch (err) {
         console.error(`DELETE KEY ROUTER ERROR : ${err}`);
     }
